intstr: add pointer constructors
This is a common idiom, useful inside k/k and outside; let's make it
official. Two functions are added, PointerFromInt32 and
PointerFromString, instead of a generic wrapper which would require
two calls in most cases.

An alternative would be to add a wrapper in the pointer functions, but
I think it makes sense to keep this self-contained.

Signed-off-by: Stephen Kitt <skitt@redhat.com>