Handle extensions and subresources generically
This requires generic functions since receivers can't be parameterised
beyond their struct's type parameters. Care must be taken when
handling subresources: their return type is the type of the
subresource, not the parent type, except for the status subresource
which uses the parent type.

delete and watch are explicitly not supported on subresources, so they
are dropped entirely.

Signed-off-by: Stephen Kitt <skitt@redhat.com>